+++ Daemon.cpp
@@ -5,8 +5,8 @@
 #include "Log.h"
 #include "base64.h"
 #include "version.h"
-#include "Transports.h"
-#include "NTCPSession.h"
+#include "transport/Transports.h"
+#include "transport/NTCPSession.h"
 #include "RouterInfo.h"
 #include "RouterContext.h"
 #include "Tunnel.h"
