+++ HTTPServer.cpp
@@ -5,7 +5,7 @@
 #include "Log.h"
 #include "Tunnel.h"
 #include "TransitTunnel.h"
-#include "Transports.h"
+#include "transport/Transports.h"
 #include "NetDb.h"
 #include "I2PEndian.h"
 #include "Streaming.h"
