+++ I2NPProtocol.cpp
@@ -8,7 +8,7 @@
 #include "NetDb.h"
 #include "Tunnel.h"
 #include "base64.h"
-#include "Transports.h"
+#include "transport/Transports.h"
 #include "Garlic.h"
 #include "I2NPProtocol.h"
