+++ Makefile
@@ -28,6 +28,7 @@ tests: mk_build_test_dir $(TESTS)
 
 mk_build_dir:
 	mkdir -p obj
+	mkdir -p obj/transport
 mk_build_test_dir:
 	mkdir -p obj/tests
 
@@ -42,12 +43,14 @@ api: $(SHLIB)
 
 deps:
 	@mkdir -p obj
+	@mkdir -p obj/transport
 	@mkdir -p obj/tests
 	$(CXX) $(CXXFLAGS) $(NEEDED_CXXFLAGS) -MM *.cpp > $(DEPS)
 	@sed -i -e '/\.o:/ s/^/obj\//' $(DEPS)
 
 obj/%.o : %.cpp
 	@mkdir -p obj
+	@mkdir -p obj/transport
 	@mkdir -p obj/tests
 	$(CXX) $(CXXFLAGS) $(NEEDED_CXXFLAGS) $(INCFLAGS) $(CPU_FLAGS) -c -o $@ $<
