+++ Makefile.bsd
@@ -7,7 +7,7 @@ CXXFLAGS = -O2
 ## -std=c++11. If you want to remove this variable please do so in a way that allows setting
 ## custom FLAGS to work at build-time.
 NEEDED_CXXFLAGS = -std=c++11
-INCFLAGS = -I/usr/include/ -I/usr/local/include/
+INCFLAGS = -I/usr/include/ -I/usr/local/include/ -I.
 LDFLAGS = -Wl,-rpath,/usr/local/lib -L/usr/local/lib
 LDLIBS = -lcryptopp -lboost_system -lboost_date_time -lboost_filesystem -lboost_regex -lboost_program_options -lpthread
 LDTESTLIBS = -lboost_unit_test_framework
