+++ Makefile.linux
@@ -1,5 +1,5 @@
 CXXFLAGS = -g -Wall
-INCFLAGS =
+INCFLAGS = -I.
 
 ## NOTE: The NEEDED_CXXFLAGS are here so that custom CXXFLAGS can be specified at build time
 ## **without** overwriting the CXXFLAGS which we need in order to build.
