+++ TransitTunnel.cpp
@@ -4,7 +4,7 @@
 #include "RouterContext.h"
 #include "I2NPProtocol.h"
 #include "Tunnel.h"
-#include "Transports.h"
+#include "transport/Transports.h"
 #include "TransitTunnel.h"
 
 namespace i2p
