+++ TunnelGateway.cpp
@@ -3,7 +3,7 @@
 #include <cryptopp/sha.h>
 #include "Log.h"
 #include "RouterContext.h"
-#include "Transports.h"
+#include "transport/Transports.h"
 #include "TunnelGateway.h"
 
 namespace i2p
