+++ filelist.mk
@@ -1,10 +1,11 @@
 COMMON_SRC = \
+  transport/NTCPSession.cpp transport/SSU.cpp transport/SSUSession.cpp \
+  transport/SSUData.cpp transport/Transports.cpp \
   CryptoConst.cpp Datagram.cpp Garlic.cpp I2NPProtocol.cpp LeaseSet.cpp \
-  Log.cpp NTCPSession.cpp NetDb.cpp NetDbRequests.cpp Profiling.cpp \
-  Reseed.cpp RouterContext.cpp RouterInfo.cpp Signature.cpp SSU.cpp \
-  SSUSession.cpp SSUData.cpp Streaming.cpp Identity.cpp TransitTunnel.cpp \
-  Transports.cpp Tunnel.cpp TunnelEndpoint.cpp TunnelPool.cpp TunnelGateway.cpp \
-  Destination.cpp UPnP.cpp util.cpp aes.cpp base64.cpp
+  Log.cpp NetDb.cpp NetDbRequests.cpp Profiling.cpp Reseed.cpp \
+  RouterContext.cpp RouterInfo.cpp Signature.cpp Streaming.cpp Identity.cpp \
+  TransitTunnel.cpp Tunnel.cpp TunnelEndpoint.cpp TunnelPool.cpp \
+  TunnelGateway.cpp Destination.cpp UPnP.cpp util.cpp aes.cpp base64.cpp
 
 
 ifeq ($(UNAME),Darwin)
